requirements.txt
asn1crypto==1.0.1
autopep8==1.4.4
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.1.5
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
passlib==1.7.1
pycodestyle==2.5.0
pycparser==2.19
pymongo==3.9.0
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
WTForms==2.2.1
